+++ src/PSTree/PSTreeDirectory.cs
@@ -1,16 +1,13 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using PSTree.Extensions;
 using PSTree.Style;
 
 namespace PSTree;
 
 public sealed class PSTreeDirectory : PSTreeFileSystemInfo<DirectoryInfo>
 {
-    private string[]? _parents;
-
-    internal string[] Parents { get => _parents ??= GetParents(FullName); }
-
     public DirectoryInfo Parent => Instance.Parent;
 
     public int ItemCount { get; internal set; }
@@ -25,7 +22,9 @@ public sealed class PSTreeDirectory : PSTreeFileSystemInfo<DirectoryInfo>
     private PSTreeDirectory(
         DirectoryInfo dir, string hierarchy, string source)
         : base(dir, hierarchy, source)
-    { }
+    {
+        ShouldInclude = true;
+    }
 
     public IEnumerable<FileInfo> EnumerateFiles() =>
         Instance.EnumerateFiles();
@@ -36,24 +35,14 @@ public sealed class PSTreeDirectory : PSTreeFileSystemInfo<DirectoryInfo>
     public IEnumerable<FileSystemInfo> EnumerateFileSystemInfos() =>
         Instance.EnumerateFileSystemInfos();
 
-    private static string[] GetParents(string path)
-    {
-        int index = -1;
-        List<string> parents = [];
-
-        while ((index = path.IndexOf(Path.DirectorySeparatorChar, index + 1)) != -1)
-        {
-            parents.Add(path.Substring(0, index));
-        }
-
-        return [.. parents];
-    }
-
-    internal IOrderedEnumerable<FileSystemInfo> GetSortedEnumerable(PSTreeComparer comparer) =>
+    internal IOrderedEnumerable<FileSystemInfo> GetSortedEnumerable(TreeComparer comparer) =>
         Instance
             .EnumerateFileSystemInfos()
             .OrderBy(static e => e is DirectoryInfo)
             .ThenBy(static e => e, comparer);
+
+    internal static PSTreeDirectory Create(string path) =>
+        Create(new DirectoryInfo(path), path);
 
     internal static PSTreeDirectory Create(DirectoryInfo dir, string source)
     {
@@ -65,5 +54,45 @@ public sealed class PSTreeDirectory : PSTreeFileSystemInfo<DirectoryInfo>
     {
         string styled = TreeStyle.Instance.GetColoredName(dir).Indent(depth);
         return new PSTreeDirectory(dir, styled, source, depth);
+    }
+
+    internal PSTreeDirectory AddParent(PSTreeDirectory parent)
+    {
+        ParentNode = parent;
+        return this;
+    }
+
+    internal void IndexCount(int count)
+    {
+        ItemCount = count;
+        TotalItemCount = count;
+
+        for (PSTreeDirectory? i = ParentNode; i is not null; i = i.ParentNode)
+        {
+            i.TotalItemCount += count;
+        }
+    }
+
+    internal void IndexLength(long length)
+    {
+        for (PSTreeDirectory? i = ParentNode; i is not null; i = i.ParentNode)
+        {
+            i.Length += length;
+        }
+    }
+
+    internal void SetIncludeFlag()
+    {
+        ShouldInclude = true;
+
+        for (PSTreeDirectory? i = ParentNode; i is not null; i = i.ParentNode)
+        {
+            if (i.ShouldInclude)
+            {
+                return;
+            }
+
+            i.ShouldInclude = true;
+        }
     }
 }