+++ src/PSTree/PSTreeFile.cs
@@ -1,4 +1,5 @@
 using System.IO;
+using PSTree.Extensions;
 using PSTree.Style;
 
 namespace PSTree;
@@ -10,14 +11,20 @@ public sealed class PSTreeFile : PSTreeFileSystemInfo<FileInfo>
     public string DirectoryName => Instance.DirectoryName;
 
     private PSTreeFile(
-        FileInfo file, string hierarchy, string source, int depth)
-        : base(file, hierarchy, source, depth) =>
+        FileInfo file, string hierarchy, string source)
+        : base(file, hierarchy, source)
+    {
         Length = file.Length;
+        ShouldInclude = true;
+    }
 
     private PSTreeFile(
-        FileInfo file, string hierarchy, string source)
-        : base(file, hierarchy, source) =>
+        FileInfo file, string hierarchy, string source, int depth)
+        : base(file, hierarchy, source, depth)
+    {
         Length = file.Length;
+        ShouldInclude = true;
+    }
 
     internal static PSTreeFile Create(FileInfo file, string source)
     {
@@ -29,5 +36,21 @@ public sealed class PSTreeFile : PSTreeFileSystemInfo<FileInfo>
     {
         string styled = TreeStyle.Instance.GetColoredName(file).Indent(depth);
         return new PSTreeFile(file, styled, source, depth);
+    }
+
+    internal PSTreeFile AddParent(PSTreeDirectory parent)
+    {
+        ParentNode = parent;
+        return this;
+    }
+
+    internal PSTreeFile SetIncludeFlagIf(bool condition)
+    {
+        if (condition)
+        {
+            ParentNode?.SetIncludeFlag();
+        }
+
+        return this;
     }
 }