+++ src/PSTree/PSTreeFileSystemInfo.cs
@@ -2,6 +2,10 @@ namespace PSTree;
 
 public abstract class PSTreeFileSystemInfo(string hierarchy, string source)
 {
+    internal PSTreeDirectory? ParentNode { get; set; }
+
+    internal bool ShouldInclude { get; set; }
+
     internal string Source { get; set; } = source;
 
     public int Depth { get; protected set; }