+++ src/PSTree/Style/Extension.cs
@@ -3,6 +3,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Management.Automation;
 using System.Text;
+using PSTree.Extensions;
 
 namespace PSTree.Style;
 
@@ -47,7 +48,7 @@ public sealed class Extension
     public override string ToString()
     {
         StringBuilder builder = new(_extension.Count);
-        int len = _extension.Keys.Max(e => e.Length);
+        int len = _extension.Keys.Max(static e => e.Length);
         foreach (KeyValuePair<string, string> pair in _extension)
         {
             builder